update lai test thong ke theo ngay

+++ Order/OrdersUI.java
@@ -22,7 +22,7 @@ public class OrdersUI {
             Customers.addCustomer(customer);
         }
 
-        int orderID = Orders.getOrders().size() + 1;
+        int orderID = Orders.generateID();
         Order newOrder = new Order(orderID, customer);
 
         boolean tiepTuc;
@@ -67,11 +67,13 @@ public class OrdersUI {
                 System.out.println("│2. Them vao hang doi                       │");
                 System.out.println("│3. Huy thiet lap                           │");
                 System.out.println("└───────────────────────────────────────────┘");
-                System.out.print("Nhap lua chon (1-3): ");
+                System.out.print("Nhap lua chon: ");
                 int choice = scanner.nextInt();
                 scanner.nextLine();
                 switch (choice) {
                     case 1:
+                        if (!payment(newOrder, customer))
+                            break;
                         newOrder.setOrderStatus(true);
                         Orders.addOrder(newOrder);
                         customer.addOrder(newOrder);
@@ -94,7 +96,89 @@ public class OrdersUI {
                         System.out.println("Lua chon khong hop le. Vui long chon lai.");
                 }
             }
-        } else System.out.println("Don hang rong. Khong them don hang.");
+        } else
+            System.out.println("Don hang rong. Khong them don hang.");
+    }
+
+    public static boolean payment(Order order, Customer customer) {
+        boolean validChoice = false;
+        while (!validChoice) {
+            System.out.println("┌───────────────────────────────────────────┐");
+            System.out.println("│        Chon phuong thuc thanh toan        │");
+            System.out.println("├───────────────────────────────────────────┤");
+            System.out.println("│1. Momo                                    │");
+            System.out.println("│2. Tien mat                                │");
+            System.out.println("│3. Quay lai                                │");
+            System.out.println("└───────────────────────────────────────────┘");
+            System.out.print("Nhap lua chon: ");
+            int choice = scanner.nextInt();
+            scanner.nextLine();
+            switch (choice) {
+                case 1:
+                    System.out.print("Nhap so dien thoai Momo: ");
+                    String momoPhone = scanner.nextLine();
+                    if (momoPhone.equals(customer.getPhone())) {
+                        System.out.print("Xac nhan thanh toan qua Momo? (y/n): ");
+                        String confirm = scanner.nextLine();
+                        if (confirm.equalsIgnoreCase("y")) {
+                            validChoice = true;
+                            return true;
+                        }
+                    } else {
+                        System.out.println("So dien thoai khong khop voi khach hang.");
+                    }
+                    break;
+                case 2:
+                    double totalAmount = order.calculateTotal();
+                    order.displayOrder();
+                    boolean paymentSuccess = false;
+                    while (!paymentSuccess) {
+                        System.out.print("Nhap so tien khach dua (boi so cua 1.000 VND): ");
+                        double amountGiven = scanner.nextDouble();
+                        scanner.nextLine();
+                        if (amountGiven >= totalAmount) {
+                            double change = amountGiven - totalAmount; // -2s
+                            System.out.println("┌───────────────────────────────────────────┐");
+                            System.out.printf("│Tien khach dua:  %-15s%-7s VND│\n", ' ',displayFormat.formatPrice(amountGiven));
+                            System.out.printf("│Tien du cua khach: %-13s%s VND│\n", ' ',displayFormat.formatPrice(change));
+                            System.out.println("├───────────────────────────────────────────┤");
+                            System.out.println("│           Thanh toan thanh cong           │");
+                            System.out.println("└───────────────────────────────────────────┘");
+                            boolean confirm = false;
+                            while (!confirm) {
+                                System.out.println("┌───────────────────────────────────────────┐");
+                                System.out.println("│1. Hoan Thanh                              │");
+                                System.out.println("└───────────────────────────────────────────┘");
+                                System.out.print("Chon 1 de hoan thanh va quay lai: ");
+                                int choice2 = scanner.nextInt();
+                                scanner.nextLine();
+                                switch (choice2) {
+                                    case 1:
+                                        confirm = true;
+                                        break;
+                                    default:
+                                        System.out.println("Lua chon khong hop le.");
+                                        break;
+                                }
+                            }
+                            paymentSuccess = true;
+                            validChoice = true;
+                            return true;
+                        } else {
+                            System.out.println("So tien khong du. Vui long nhap lai.");
+                        }
+                    }
+                    break;
+                case 3:
+                    System.out.println("Quay lai.");
+                    validChoice = true;
+                    return false;
+                default:
+                    System.out.println("Lua chon khong hop le. Vui long chon lai.");
+                    break;
+            }
+        }
+        return false;
     }
 
     // Xem danh sách đơn hàng
@@ -119,7 +203,12 @@ public class OrdersUI {
                     System.out.print("Nhap ma don hang can xoa: ");
                     int orderID = scanner.nextInt();
                     scanner.nextLine();
-                    Orders.deleteOrder(orderID);
+                    System.out.print("Ban co chan chan mua xoa don hang? (y/n): ");
+                    String confirm = scanner.nextLine();
+                    if (confirm.equalsIgnoreCase("y")) {
+                        complete = true;
+                        Orders.deleteOrder(orderID);
+                    }
                     break;
                 case 3:
                     return;
@@ -131,7 +220,7 @@ public class OrdersUI {
 
     // Chỉnh sửa đơn hàng
     public static void editOrder() {
-        System.out.print("Nhap ma don hang can sua: ");
+        System.out.print("Nhap ma don hang ban muon chon: ");
         int orderID = scanner.nextInt();
         scanner.nextLine();
         Orders.edit(orderID); // Truy cập phương thức tĩnh
